requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.1
python-multipart>=0.0.9
slowapi>=0.1.9

# Parsing
pdfplumber>=0.11.0
python-docx>=1.1.0

# Embeddings
sentence-transformers>=3.0.0

# Vector store
faiss-cpu>=1.8.0
numpy>=1.26.0

# LLM
openai>=1.30.0